Fix perf run trypes mismatches
